Clean up use of size_t vs. uint_32 vs. int.
git-svn-id: svn://svn.savannah.nongnu.org/pngpp/trunk@138 c52bbda5-c92d-0410-b1d5-fe105dfd41d1

+++ info.hpp
@@ -105,7 +105,7 @@ namespace png
                 {
                     png_set_PLTE(m_png, m_info,
                                  const_cast< color* >(& m_palette[0]),
-                                 m_palette.size());
+                                 (int) m_palette.size());
                 }
                 if (! m_tRNS.empty())
                 {
